Make team IDs int64 and add flag collector
Convert TeamID fields to int64 across server and client, adding
casts where needed. Move API request types into pkg/models and
update client usage. Add a buffered FlagCollector singleton to batch
DB writes with Start/Stop/Flush methods. Embed DB schema and update
sqlc config; add mapper helpers and CountFlags query. Misc:
change DSN to cookiefarm.db, adjust route handlers, and update some
logging and module/sum entries.

+++ cookiefarm/server/websockets/event.go
@@ -51,7 +51,7 @@ func FlagHandler(event Event, client *Client) error {
 	logger.Log.Info().
 		Int("client", client.Number).
 		Str("flag", flag.FlagCode).
-		Uint16("team", flag.TeamID).
+		Int64("team", flag.TeamID).
 		Str("service name", flag.ServiceName).
 		Uint16("port service", flag.PortService).
 		Msg("Flag received and sent to DB")
